Correct Quicksort
There were two problems with Quicksort:

1. while(l < r) (strict non equality) could lead to a situation when
the left and right moving pointers stopped their movement at the same
position (this was the case if there was a swap for r - l == 2, just
before the stop). Sometimes it led to pivot element being swapped with
an incorrect element at the end (with index l, for example, instead of
r = l -1).

2. In the previous version, pivot was always swapped with the right
pointer. It is incorrect because the right pointer can also, just like
the left one, leave from the array. Now pivot is swapped with the
*nearest* pointer, which is guaranteed to be in range.