Replace tree-like ascending with linear scan
The point is that in a perfect binary tree there are precisely as many
nodes as many times one needs to call GCD(x, y) function in a simple
linear scan.

So it is wise to switch to a simple linear scan, which makes it MUCH
easier to get rid from zeros before calling GetGCDBase, which is
promised to be passed positive arguments only.