Rewrite CombineWith mechanism
The CombineWith mechanism now works as follows:

1. Each matrix layout defines two CombineWith methods: one takes a
layout of the same type, the other takes a layout interface and is
hidden behind explicit interface implementation.

2. The method which takes a concrete layout can automatically take any
more specific layouts via an implicit cast. This method uses information
about the concrete layout for optimization.

3. All other cases (layouts whose level of generality is either higher
or unknown) must be combined resorting to an explicit interface method.